Remove href from back link to prevent URL change;
When opening the remote on OS X dashboard, clicking the back button will
cause the index page to be opened in Safari. This is a result of the URL
changing which in turn is caused by the href attribute. This attribute
is however not needed and can thus be removed.