Add button and remote labels to config to support LIRC namespaces;
Earlier implementation showed remote and button names in the form they
were in lircd.conf. In LIRC configuration guide
(http://www.lirc.org/html/configuration-guide.html#appendix-4) it is
advised that standard key symbols are used in the configuration file.
Because this is not the best format for readability it is useful to be
able to specify a key symbol to label mapping in the configuration file.

An example on how to use the labels can be found in
test/fixtures/config.json.